App/requirements.txt
flask
flask_restful
requests
PyGithub
jwt
pythainlp
sklearn
psycopg2-binary